Always load deep objects; use routines to convert
For objects that were updated, this groups and family curation pages read in flattened objects to make writing easier, but then needed to load articles as a separate object. With this change, objects are always loaded fleshed out, and then a flattened copy is made for writing. This simplifies the code, and reduces network traffic a tad.