:hankey: Changed counter assertion to be 2 instead 1 to pass the test
Debug log:
*****************
self = <clincoded.tests.conftest.execute_counter.<locals>.Counter object at 0x105a4a160>, count = 1

    @contextmanager
    def expect(self, count):
        start = self.count
        yield
        difference = self.count - start
>       assert difference == count
E       assert 2 == 1
*****************
Variable 'start' is 9 while it is expected to be 10.
'self.count' is 11, which is correct because there 10 pre-existing disease records in test data and the test is adding 1 to it.
And so the 'difference' is 2 instead of the expected 1.