:bug: Fixed bug in which changing GDM disease from ontology term to free text failed to save/update the GDM object properly
The initial attempt was to only create a new free text disease record if we are changing the GDM disease from ontology term to free text term. And if we are simply just updating a free text term by modifying the short term name, the HPO IDs and/or the definition, then there is no need to create a new record for it and we shall just do a PUT request to update the current free text disease term for the GDM. We shall revisit this issue and figure out why it is not working as expected conceptually.