#1467 :bug: Fixed bug introduced in R12. #1335 #1444 #1446 Additional changes in text descriptions and labels. See details.
In R12, I implemented a feature to disable scoring within various Experimental curation sections until the user explicitly selects a given checkbox or this checkbox has a true boolean value when the curation page is rendered with a pre-existing data record.

However, I overlooked the sections (e.g. Functional Alteration, Model Systems and the Biochemical Function 'B' option) that do not have the similar checkbox described above. It is worth noting that this bug does not affect these sections when the user starts a curation from scratch.