Removing my workaround for the failed Travis CI builds to avoid conflicts when merging with the latest 'dev' branch. See detail.
Chris took this workaround and tested it in a separate branch without the 'boto' and 'six' module version upgrades. After verifying the version upgrades were not needed, Chris submitted a PR for this workaround and I merged it to 'dev' so that the fix can be used by other active branches.