#1414 Converted this file to become the classification view-only page. See details.
Reconstructed this classification view-only page without the imported matrix and form components. It is currently using the same logic to calculate and render the GCI score matrix and it is not optimal. Ideally it should retrieve all the score data from database rather doing the calculations again on a view-only page when versioning of a given gdm has been implemented.

This view-only page is essentially a simplified version of the classification matrix and form page. Aside from the matrix, we are simply just displaying the stored classification form values (no form required).

3 new buttons have also been implemented for this view-only page:
1. View Evidence Summary button that opens up a new window showing all relevant evidence.
2. Edit Classification button that takes the use back to the classification form edit page.
3. Curation-Central button that takes the user back to curation central.