Changing disease 'OMIM_261600' to an entirely different disease
Because MonDO ontology changed the clique leader for disease 'phenylketonuria' from 'OMIM:261600' to 'Orphanet:716', and then from 'Orphanet:716' to 'MONDO:0009861' in the upcoming release, there are interpretations that had been associated with this same disease having 2 different disease IDs.

As a result of migrating all diseases to using MonDO IDs, both of the 'Orphanet:761' and 'OMIM:261600' will be changed to the same 'MONDO:0009861' ID, and therefore, we will end up with duplicate disease entries. An error will also likely occur when the upgrade script runs during the build process.

The workaround is to change all interpretations from be associating with 'OMIM:261600' to be associating with 'Orphanet:761' (via UUIDs) during the data upgrade process. And 'OMIM:261600' can be changed to an entirely different disease while 'Orphanet:761' can be migrated to the mapped 'MONDO:0009861' ID.