Removing 'DOID:2481' from the MonDO ID migration
'DOID:2481' has been incorrectly mapped to 'MONDO:0016021' in MonDO ontology ('infantile epileptic encephalopathy' is NOT the same as 'early infantile epileptic encephalopathy'). So we are holding off on migrating this ID until clarification is received.

For reference, below is the removed code from the upgrade script:
if value['diseaseId'] == 'DOID_2481':
    value['diseaseId'] = 'MONDO_0016021'
    value['term'] = 'early infantile epileptic encephalopathy'
    value['description'] = 'Early infantile epileptic encephalopathy (EIEE), or Ohtahara syndrome, is one of the most severe forms of age-related epileptic encephalopathies, characterized by the onset of tonic spasms within the first 3 months of life that can be generalized or lateralized, independent of the sleep cycle and that can occur hundreds of times per day, leading to psychomotor impairment and death.'