Updated the rendering of editable interpretations in the static header
Due to the implementation of affiliation, the affiliated interpretation will be rendered as an editable item rather than the individual's interpretation if the user is logged in with an affiliation. If the user is logged in as an individual, then we will  render the individual's interpretation as an editable item.