Updated the behaviors of the variant interpretations table and GDM records table on the Dashboard. See details.
Per discussions with the curators, we do not want to show the affiliated evidence as separate tables organized by types.

Going forward, when a curator signs in as a member of an affiliation, on the Dashboard page, we show the affiliated variant interpretations as well as the affiliated GDM records. The affiliated GDM records include affiliated GDMs, affiliated annotations, affiliated evidence, affiliated scores and affiliated classifications.

And if the curator signs in without an affiliation, we just show the variant interpretations and GDM records pertinent to the curator.