:bug: It broke the interpretation evaluation and prevented it from being saved
The problem is that 'modified_by' is not defined in the curationHistory.json schema's 'meta.interpretation' property.

Adding the modified_by user name to the interpretation history item was meant to be a cheap workaround for the lack of the modified_by object in the history.primary object after an evaluation is saved.

Will need to revisit this for a proper implementation.