Allow the 'Explanation' text field to be enabled for use when scoring even if the curator is not changing from default score
The 'Explanation' text field is now enabled to all 'Score Status', including Mode of Inheritance that are not AD, AR and X-Linked.

The exception is that the 'Explanation' text field remains disabled when the 'Score Status' is 'Score' and there is no 'Case Information Type' being specified.

Various micro-interactions in the UI have been updated given the new specs.