UI changes in the Proband score panel
In the top yellow note box in the Proband Score panel, links have been added to allow users to curate variant's gene impact as long as one or more variants being associated with the Proband.

Clicking on the variant link will open a new window/tab so that the user will not lose the potentially unfinished Proband curation data in the form.