Add handling to contruct variant title consisting of canonical transcript, gene symbol and protein change
For CA IDs that don't have equivalent ClinVar title in the request response, we ping the Ensembl VEP Rest API to get their transcripts and pick out the canonical transcript. Once the canonical transcript is identified, we use it to pick out the gene symbol and protein change from the temp allele arrays for constructing the variant title in the variant object. After that, we delete the temp allele arrays.