:bug: #1687 #1690 Fixed bug in which duplicate HPO terms were rendered in different proband table rows
The <HpoTerms> rendering component's own state does not work properly when each evidence has multiple HPO IDs to fetch data for. As a result, moving the fetching handler to this entry page of the evidence summary.

The strategy is to create a HPO id/term dictionary of aggregated HPO IDs for each evidence type (e.g. case-level, segregation, case-control). Then pass it to the child components to handle the mappings when rendering the HPO terms.

Also fixed a bug related to #1690 in which the parsing is not correctly handling family/segregation evidence that have a mixture of scored probands and non-scored probands andnon-proband individuals.